refactor: remove MultipleAmountField component and update related imports in forms
