+++ apps/app/components/fields/index.ts
@@ -8,7 +8,6 @@ export * from './email-field'
 export * from './file-upload-field'
 export * from './icon-field'
 export * from './ignore-field'
-export * from './multiple-amount-field'
 export * from './note-field'
 export * from './password-field'
 export * from './select-field'
